+++ lib/src/selected/mp_selected_point.dart
@@ -10,12 +10,20 @@ class MPSelectedPoint extends MPSelectedElement {
   void _createClone(THPoint originalPoint) {
     final SplayTreeMap<THCommandOptionType, THCommandOption> optionsMapClone =
         SplayTreeMap<THCommandOptionType, THCommandOption>();
+    final SplayTreeMap<String, THAttrCommandOption> attrOptionsMapClone =
+        SplayTreeMap<String, THAttrCommandOption>();
 
     originalPoint.optionsMap.forEach((key, value) {
       optionsMapClone[key] = value.copyWith();
     });
+    originalPoint.attrOptionsMap.forEach((key, value) {
+      attrOptionsMapClone[key] = value.copyWith();
+    });
 
-    originalPointClone = originalPoint.copyWith(optionsMap: optionsMapClone);
+    originalPointClone = originalPoint.copyWith(
+      optionsMap: optionsMapClone,
+      attrOptionsMap: attrOptionsMapClone,
+    );
   }
 
   @override